Unify Device.cpp with WSLDevice.cpp and remove WSLDevice.cpp
